requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==3.3.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==4.6.1
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
openpyxl==3.0.8
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.2
parso==0.8.2
pickleshare==0.7.5
pip==21.1.3
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
pywin32==300
pywinpty==1.1.3
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.25.1
Send2Trash==1.7.1
setuptools==49.6.0.post20210108
six==1.16.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==2.0.1
zipp==3.5.0
